refactor: make brainfuck-vm a library and refactor binary
