+++ crates/brainfuck-vm/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod compiler;
+pub mod instruction;
+pub mod machine;
+pub mod registers;
+pub mod test_helper;
